manager/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
fastmcp==2.14.5
httpx>=0.28.0
pyyaml>=6.0.2
python-dotenv>=1.0.0
cryptography>=43.0.0
GitPython>=3.1.0
